Added support for custom font on title and detail
It's in the ConfigurationOptions now. These Strings specify the Asset
name of the TTF font you want to use. The Asset must be put in the main
project (not THIS project if you're including it as a library) in
assets/... on the same level as the src directory. Example: for file
location /git/projectname/assets/fonts/myfont.ttf  use string
"fonts/myfont.ttf"