Fix incorrect start position of showcase animation
ShowcaseView#setShowcasePosition adjusts the given coordinates
for the location of the ShowcaseView inside the window.
When ObjectAnimator calls getShowcaseY to get the starting value
for the animation, the already adjusted value was returned,
thus basically incorrectly applying the adjustment twice when
the ObjectAnimator sets that value at the start of the animation.

As a fix, hide the fact that the fields showcaseX and showcaseY
are adjusted by returning the un-adjusted values from their getters.